ZunMemory::Alloc - u32->size_t

+++ src/ZunMemory.hpp
@@ -6,7 +6,7 @@ namespace th06
 {
 namespace ZunMemory
 {
-inline void *Alloc(u32 size)
+inline void *Alloc(size_t size)
 {
     return malloc(size);
 }
